Changed from global to local movement tracking. This significantly reduces the number of actions that need to be logged.
Made lookat disable on replay. This will significantly reduce the number of actions that need to be logged as well as remove jerky movement of all text by removing the tug of war between the replay and the behavior. This is not yet tested.

Objects are sometimes jumping instead of smooth lerping. Issue needs to be investigated.

Objects on replay are extremely large and incorrectly positioned. Issue might have to do with the way size is handeled on load. Issue needs investigating.