MAde spawns happen slightly early and destroys happen slightly late to prevent other logs from null referencing.
Should still happen at about the right time, but might not be enough. If issue persists may try changing the implementation of clock to only update at the beggining of a frame so that all actions in the frame happen at the same time. This would force spawns and destroys to happen first or last respectively.