fixed dumb bug in spawn that caused it to deactivate.
fixed virtual button so that it will actually activate when pressed in editor.

simplified Button class and made it clear that editorpress will not work in build.

added grab simulation. Test passes, but can't tell if properly resolved.

Getting a null reference on code that shouldn't really be doing anything when it null references. -- load code null referencing at end of replay.